backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
pymongo
python-multipart